The build report now shows the student unit tests results (fix #10); The plugin doesn't overwrite the AUTHORS.txt file anymore

+++ src/main/kotlin/org/dropProject/dropProjectPlugin/AuthorsFile.kt
@@ -10,8 +10,13 @@ class AuthorsFile(private val students: ArrayList<User>) {
 
     fun make(projectDirectory: String?, silentMode: Boolean, e: AnActionEvent) {
 
-
         val authors = File("$projectDirectory${File.separator}AUTHORS.txt")
+
+        if (authors.exists()) {
+            if (!silentMode)
+                DefaultNotification.notify(e.project, "AUTHORS file already exists so I will not modify it")
+            return
+        }
 
         if (authors.delete()) {
             if (!silentMode)
